# gravitee-apim-rest-api/gravitee-apim-rest-api-management-v4/gravitee-apim-rest-api-management-v4-rest/src/main/resources/management-openapi-v4.yaml
openapi: 3.0.3
info:
    title: Gravitee.io APIM - Management API
    description: |-
        This is the OpenAPI specification for our new version of APIM Management API.
    contact:
        email: team-apim@graviteesource.com
    license:
        name: Apache 2.0
        url: http://www.apache.org/licenses/LICENSE-2.0.html
    version: 4.0.0-alpha.1
...
servers:
    - url: https://apim-{version}-api.team-apim.gravitee.dev/management/v4
      description: Gravitee.io APIM - Management API -v4
      variables:
          version:
              default: 4-0-x
paths:
...
    /endpoints:
        get:
            parameters:
                - $ref: "#/components/parameters/pageParam"
                - $ref: "#/components/parameters/perPageParam"
            tags:
                - Connectors
                - Endpoints
            summary: Get available endpoints on the platform
            description: |
                # FIXME add a description
            operationId: getEndpoints
            responses:
                "200":
                    $ref: "#/components/responses/EndpointsResponse"
                default:
                    $ref: "#/components/responses/Error"
    /endpoints/{endpointId}:
        parameters:
            - $ref: "#/components/parameters/endpointIdParam"
        get:
            tags:
                - Connectors
                - Endpoints
            summary: Get an endpoint
            description: |
                # FIXME add a description
            operationId: getEndpoint
            responses:
                "200":
                    description: Endpoint
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ConnectorPlugin"
                default:
                    $ref: "#/components/responses/Error"
    /endpoints/{endpointId}/schema:
        parameters:
            - $ref: "#/components/parameters/endpointIdParam"
        get:
            tags:
                - Connectors
                - Endpoints
            summary: Get an endpoint schema
            description: |
                # FIXME add a description
            operationId: getEndpointSchema
            responses:
                "200":
                    $ref: "#/components/responses/SchemaFormResponse"
                default:
                    $ref: "#/components/responses/Error"
    /endpoints/{endpointId}/documentation:
        parameters:
            - $ref: "#/components/parameters/endpointIdParam"
        get:
            tags:
                - Connectors
                - Endpoints
            summary: Get an endpoint documentation
            description: |
                # FIXME add a description
            operationId: getEndpointDocumentation
            responses:
                "200":
                    description: Documentation of the endpoint
                    content:
                        text/plain:
                            schema:
                                type: string
                default:
                    $ref: "#/components/responses/Error"
    /endpoints/{endpointId}/endpointGroupConfiguration:
        parameters:
            - $ref: "#/components/parameters/endpointIdParam"
        get:
            tags:
                - Connectors
                - Endpoints
            summary: Get an endpoint group configuration schema
            description: |
                # FIXME add a description
            operationId: getEndpointGroupSchema
            responses:
                "200":
                    $ref: "#/components/responses/SchemaFormResponse"
                default:
                    $ref: "#/components/responses/Error"
...
components:
    schemas:
...
        ApiType:
            type: string
            description: API's type.
            example: async
            enum:
                - async
                - sync
...
        ConnectorFeature:
            type: string
            description: A feature used by a connector
            example: limit
            enum:
                - limit
                - resume
                - dlq
        ConnectorMode:
            type: string
            description: A mode used by a connector
            example: subscribe
            enum:
                - subscribe
                - publish
                - request_response
        ConnectorPlugin:
            type: object
            allOf:
                - $ref: "#/components/schemas/PlatformPlugin"
                - type: object
                  properties:
                      supportedApiType:
                          $ref: "#/components/schemas/ApiType"
                      supportedModes:
                          type: array
                          items:
                              $ref: "#/components/schemas/ConnectorMode"
                          uniqueItems: true
                      supportedQos:
                          type: array
                          items:
                              $ref: "#/components/schemas/QoS"
                          uniqueItems: true
                      supportedListenerType:
                          $ref: "#/components/schemas/ListenerType"
                      availableFeatures:
                          type: array
                          items:
                              $ref: "#/components/schemas/ConnectorFeature"
                          uniqueItems: true
                      schema:
                          type: string
                          description: The schema of the plugin.
                      icon:
                          type: string
                          description: The icon of the plugin.
                      subscriptionSchema:
                          type: string
                          description: The subscription schema of the plugin.
...
        Error:
            type: object
            properties:
                httpStatus:
                    type: integer
                    format: int32
                    description: The error code
                    example: 400
                message:
                    type: string
                    description: The error message
                    example: Bad request
                technicalCode:
                    type: string
                    description: A technical code to identify the error
                    example: invalid.import.definition
                parameters:
                    type: object
                    description: A map of parameters to be used in the error message
                    additionalProperties:
                        type: string
...
        Links:
            description: List of links for pagination
            properties:
                self:
                    type: string
                    description: Link to current resource
                first:
                    type: string
                    description: In a paginated response, link to the first page
                last:
                    type: string
                    description: In a paginated response, link to the last page
                prev:
                    type: string
                    description: In a paginated response, link to the previous page. Maybe null if current is the first page
                next:
                    type: string
                    description: In a paginated response, link to the next page. Maybe null if current is the last page
...
        ListenerType:
            type: string
            description: Listener type.
            example: http
            enum:
                - http
                - subscription
                - tcp
...
        MetadataResponse:
            description: Generic object to handle additional information about an entity. Can also be used for pagination data.
            type: object
...
        PlatformPlugin:
            type: object
            properties:
                id:
                    type: string
                    description: Plugin's uuid.
                    example: 00f8c9e7-78fc-4907-b8c9-e778fc790750
                name:
                    type: string
                    description: Plugin's name.
                    example: My Plugin
                    minLength: 1
                description:
                    type: string
                    description: Plugin's description.
                    example: I can use many characters to describe this plugin.
                    minLength: 1
                category:
                    type: string
                    description: Plugin's category.
                    example: policy
                    minLength: 1
                version:
                    type: string
                    description: Plugin's version.
                    example: 1.0.0
                    minLength: 1
        QoS:
            type: string
            description: Type of the quality of service.
            example: none
            enum:
                - none
                - auto
                - at-most-once
                - at-least-once
...
    parameters:
...
        endpointIdParam:
            name: endpointId
            in: path
            required: true
            description: Id of an endpoint.
            schema:
                type: string
...
        pageParam:
            name: page
            in: query
            required: false
            description: The page number for pagination.
            schema:
                type: integer
                default: 1
        perPageParam:
            name: perPage
            in: query
            required: false
            description: |
                The number of items per page for pagination.
            schema:
                type: integer
                default: 10
    responses:
...
        EndpointsResponse:
            description: Page of endpoints
            content:
                application/json:
                    schema:
                        properties:
                            data:
                                description: List of Endpoints.
                                type: array
                                items:
                                    $ref: "#/components/schemas/ConnectorPlugin"
                            metadata:
                                $ref: "#/components/schemas/MetadataResponse"
                            links:
                                $ref: "#/components/schemas/Links"
...
        SchemaFormResponse:
            description: Schema form of a plugin
            content:
                application/json:
                    schema:
                        type: object
        Error:
            description: Generic error response
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/Error"